SPEC/SOFTWARE/queue_monitor/requirements.txt
opencv-python>=4.5.0
ultralytics>=8.0.0
flask>=2.0.0
flask-cors>=3.0.0
numpy>=1.20.0
pyyaml>=5.4.0
pillow>=8.0.0
RPi.GPIO>=0.7.0
